Initialize scripts (#3)
* Add scripts

* Updates

* Add RKE2 windows sample

* Update RKE2 windows

* Update

* Fix RKE2 CoreDNS issue with Windows node

* Fix typo

* Add Ingress config

* Add investigation results on windows container error with calico

* Add docker_install_apt

* Add shell scripting doc

* Add comment for CAPI

* Create project-setup.md

* Updates Helm repository for Let's Encrypt chart

The previous Helm repository `suse-lab-setup` is deprecated and replaced with `devpro` to ensure continued access to the Let's Encrypt chart.
This change updates the Helm repository URL and chart location used to install the Let's Encrypt certificate issuer.

* Refactors cluster ID and registration commands

Removes redundant functions for retrieving cluster ID and registration commands.

Simplifies the code by directly using the "return" functions instead of separate "get" functions that only echoed debug information.

* Adds host startup wait function

Introduces a function to wait for the completion of the host startup bootstrap process.

This ensures that subsequent operations requiring a fully initialized host are not executed prematurely.

* Adds Gitea Kubernetes installation script

Adds a script to facilitate Gitea installation within a Kubernetes cluster as a single pod, along with necessary configurations for Ingress and certificate management.
Updates the README files to reflect the latest changes and script additions.

* Improve wait in gitea_kubernetes_install_singlepod